읽어올 때 문제가 있는 것 같음
확인 부탁

+++ Stardew_Valley/Stardew_Valley/Framework/Utility/Timer.cpp
@@ -9,20 +9,20 @@ Timer::Timer()
 
 	QueryPerformanceCounter((LARGE_INTEGER*)&_lastTime);
 
-	_timeScale = 1.0 / (double)_periodFrequency;
+	_timeScale = 1.0f / (float)_periodFrequency;
 }
 
 void Timer::Update()
 {
 	QueryPerformanceCounter((LARGE_INTEGER*)&_curTime);
-	_deltaTime = (double)(_curTime - _lastTime) * _timeScale;
+	_deltaTime = (float)(_curTime - _lastTime) * _timeScale;
 
 	if (_lockFPS != 0)
 	{
-		while (_deltaTime < (1.0 / _lockFPS))
+		while (_deltaTime < (1.0f / _lockFPS))
 		{
 			QueryPerformanceCounter((LARGE_INTEGER*)&_curTime);
-			_deltaTime = (double)(_curTime - _lastTime) * _timeScale;
+			_deltaTime = (float)(_curTime - _lastTime) * _timeScale;
 		}
 	}
