+++ Stardew_Valley/Stardew_Valley/Framework/Utility/Timer.h
@@ -30,15 +30,15 @@ public:
 	void Update();
 
 	int GetFPS() { return _frameRate; }
-	double GetDeltaTime() { return _deltaTime; }
-	double GetRunTime() { return _runTime; }
-	void LockRunTime(double val) { _lockFPS = val; }
+	float GetDeltaTime() { return _deltaTime; }
+	float GetRunTime() { return _runTime; }
+	void LockRunTime(float val) { _lockFPS = val; }
 
 private:
 	static Timer* _instance;
 
-	double _timeScale = 0.0;
-	double _deltaTime = 0.0;
+	float _timeScale = 0.0;
+	float _deltaTime = 0.0;
 
 	unsigned __int64 _curTime = 0;
 	unsigned __int64 _lastTime = 0;
@@ -47,8 +47,8 @@ private:
 	unsigned int _frameRate = 0;
 	unsigned int _frameCount = 0;
 
-	double _oneSecCount = 0.0;
-	double _runTime = 0.0;
-	double _lockFPS = 0.0;
+	float _oneSecCount = 0.0;
+	float _runTime = 0.0;
+	float _lockFPS = 0.0;
 };
